+++ src/Twilio/Rest/FlexApi/V1/InsightsQuestionnairesCategoryResource.cs
@@ -171,6 +171,130 @@ namespace Twilio.Rest.FlexApi.V1
         }
         #endif
         
+        private static Request BuildReadRequest(ReadInsightsQuestionnairesCategoryOptions options, ITwilioRestClient client)
+        {
+            
+            string path = "/v1/Insights/QM/Categories";
+
+
+            return new Request(
+                HttpMethod.Get,
+                Rest.Domain.FlexApi,
+                path,
+                queryParams: options.GetParams(),
+                headerParams: options.GetHeaderParams()
+            );
+        }
+        /// <summary> To get all the categories </summary>
+        /// <param name="options"> Read InsightsQuestionnairesCategory parameters </param>
+        /// <param name="client"> Client to make requests to Twilio </param>
+        /// <returns> A single instance of InsightsQuestionnairesCategory </returns>
+        public static ResourceSet<InsightsQuestionnairesCategoryResource> Read(ReadInsightsQuestionnairesCategoryOptions options, ITwilioRestClient client = null)
+        {
+            client = client ?? TwilioClient.GetRestClient();
+            var response = client.Request(BuildReadRequest(options, client));
+            var page = Page<InsightsQuestionnairesCategoryResource>.FromJson("categories", response.Content);
+            return new ResourceSet<InsightsQuestionnairesCategoryResource>(page, options, client);
+        }
+
+        #if !NET35
+        /// <summary> To get all the categories </summary>
+        /// <param name="options"> Read InsightsQuestionnairesCategory parameters </param>
+        /// <param name="client"> Client to make requests to Twilio </param>
+        /// <returns> Task that resolves to A single instance of InsightsQuestionnairesCategory </returns>
+        public static async System.Threading.Tasks.Task<ResourceSet<InsightsQuestionnairesCategoryResource>> ReadAsync(ReadInsightsQuestionnairesCategoryOptions options,
+                                                                                             ITwilioRestClient client = null)
+        {
+            client = client ?? TwilioClient.GetRestClient();
+            var response = await client.RequestAsync(BuildReadRequest(options, client));
+
+            var page = Page<InsightsQuestionnairesCategoryResource>.FromJson("categories", response.Content);
+            return new ResourceSet<InsightsQuestionnairesCategoryResource>(page, options, client);
+        }
+        #endif
+        /// <summary> To get all the categories </summary>
+        /// <param name="token"> The Token HTTP request header </param>
+        /// <param name="pageSize"> How many resources to return in each list page. The default is 50, and the maximum is 1000. </param>
+        /// <param name="limit"> Record limit </param>
+        /// <param name="client"> Client to make requests to Twilio </param>
+        /// <returns> A single instance of InsightsQuestionnairesCategory </returns>
+        public static ResourceSet<InsightsQuestionnairesCategoryResource> Read(
+                                                     string token = null,
+                                                     int? pageSize = null,
+                                                     long? limit = null,
+                                                     ITwilioRestClient client = null)
+        {
+            var options = new ReadInsightsQuestionnairesCategoryOptions(){ Token = token, PageSize = pageSize, Limit = limit};
+            return Read(options, client);
+        }
+
+        #if !NET35
+        /// <summary> To get all the categories </summary>
+        /// <param name="token"> The Token HTTP request header </param>
+        /// <param name="pageSize"> How many resources to return in each list page. The default is 50, and the maximum is 1000. </param>
+        /// <param name="limit"> Record limit </param>
+        /// <param name="client"> Client to make requests to Twilio </param>
+        /// <returns> Task that resolves to A single instance of InsightsQuestionnairesCategory </returns>
+        public static async System.Threading.Tasks.Task<ResourceSet<InsightsQuestionnairesCategoryResource>> ReadAsync(
+                                                                                             string token = null,
+                                                                                             int? pageSize = null,
+                                                                                             long? limit = null,
+                                                                                             ITwilioRestClient client = null)
+        {
+            var options = new ReadInsightsQuestionnairesCategoryOptions(){ Token = token, PageSize = pageSize, Limit = limit};
+            return await ReadAsync(options, client);
+        }
+        #endif
+
+        
+        /// <summary> Fetch the target page of records </summary>
+        /// <param name="targetUrl"> API-generated URL for the requested results page </param>
+        /// <param name="client"> Client to make requests to Twilio </param>
+        /// <returns> The target page of records </returns>
+        public static Page<InsightsQuestionnairesCategoryResource> GetPage(string targetUrl, ITwilioRestClient client)
+        {
+            client = client ?? TwilioClient.GetRestClient();
+
+            var request = new Request(
+                HttpMethod.Get,
+                targetUrl
+            );
+
+            var response = client.Request(request);
+            return Page<InsightsQuestionnairesCategoryResource>.FromJson("categories", response.Content);
+        }
+
+        /// <summary> Fetch the next page of records </summary>
+        /// <param name="page"> current page of records </param>
+        /// <param name="client"> Client to make requests to Twilio </param>
+        /// <returns> The next page of records </returns>
+        public static Page<InsightsQuestionnairesCategoryResource> NextPage(Page<InsightsQuestionnairesCategoryResource> page, ITwilioRestClient client)
+        {
+            var request = new Request(
+                HttpMethod.Get,
+                page.GetNextPageUrl(Rest.Domain.Api)
+            );
+
+            var response = client.Request(request);
+            return Page<InsightsQuestionnairesCategoryResource>.FromJson("categories", response.Content);
+        }
+
+        /// <summary> Fetch the previous page of records </summary>
+        /// <param name="page"> current page of records </param>
+        /// <param name="client"> Client to make requests to Twilio </param>
+        /// <returns> The previous page of records </returns>
+        public static Page<InsightsQuestionnairesCategoryResource> PreviousPage(Page<InsightsQuestionnairesCategoryResource> page, ITwilioRestClient client)
+        {
+            var request = new Request(
+                HttpMethod.Get,
+                page.GetPreviousPageUrl(Rest.Domain.Api)
+            );
+
+            var response = client.Request(request);
+            return Page<InsightsQuestionnairesCategoryResource>.FromJson("categories", response.Content);
+        }
+
+        
         private static Request BuildUpdateRequest(UpdateInsightsQuestionnairesCategoryOptions options, ITwilioRestClient client)
         {