+++ src/Twilio/Rest/Supersim/V1/SmsCommandResource.cs
@@ -246,7 +246,7 @@ namespace Twilio.Rest.Supersim.V1
         #endif
         /// <summary> Retrieve a list of SMS Commands from your account. </summary>
         /// <param name="sim"> The SID or unique name of the Sim resource that SMS Command was sent to or from. </param>
-        /// <param name="status"> The status of the SMS Command. Can be: `queued`, `sent`, `delivered`, `received` or `failed`. See the [SMS Command Status Values](https://www.twilio.com/docs/wireless/api/smscommand-resource#status-values) for a description of each. </param>
+        /// <param name="status"> The status of the SMS Command. Can be: `queued`, `sent`, `delivered`, `received` or `failed`. See the [SMS Command Status Values](https://www.twilio.com/docs/iot/supersim/api/smscommand-resource#status-values) for a description of each. </param>
         /// <param name="direction"> The direction of the SMS Command. Can be `to_sim` or `from_sim`. The value of `to_sim` is synonymous with the term `mobile terminated`, and `from_sim` is synonymous with the term `mobile originated`. </param>
         /// <param name="pageSize"> How many resources to return in each list page. The default is 50, and the maximum is 1000. </param>
         /// <param name="limit"> Record limit </param>
@@ -267,7 +267,7 @@ namespace Twilio.Rest.Supersim.V1
         #if !NET35
         /// <summary> Retrieve a list of SMS Commands from your account. </summary>
         /// <param name="sim"> The SID or unique name of the Sim resource that SMS Command was sent to or from. </param>
-        /// <param name="status"> The status of the SMS Command. Can be: `queued`, `sent`, `delivered`, `received` or `failed`. See the [SMS Command Status Values](https://www.twilio.com/docs/wireless/api/smscommand-resource#status-values) for a description of each. </param>
+        /// <param name="status"> The status of the SMS Command. Can be: `queued`, `sent`, `delivered`, `received` or `failed`. See the [SMS Command Status Values](https://www.twilio.com/docs/iot/supersim/api/smscommand-resource#status-values) for a description of each. </param>
         /// <param name="direction"> The direction of the SMS Command. Can be `to_sim` or `from_sim`. The value of `to_sim` is synonymous with the term `mobile terminated`, and `from_sim` is synonymous with the term `mobile originated`. </param>
         /// <param name="pageSize"> How many resources to return in each list page. The default is 50, and the maximum is 1000. </param>
         /// <param name="limit"> Record limit </param>